feat: per-unit options (fake touch, blinking, meters on display) + Plug-Mode LCD meter fix
Move the former global CONFIG_FLAG_FADER_TOUCH_FAKE and
CONFIG_FLAG_EMULATING_BLINKING options from a single surface-wide flag
into per-unit UNIT_FLAG_* bits (UnitConfig.unitFlags), so each hardware
unit can opt in independently. Add a new persisted-but-not-yet-wired
per-unit option "Show level meters on the display"
(UNIT_FLAG_METERS_ON_DISPLAY), plus UNIT_FLAG_SWITCH_ROWS.

Config format:
- KLINKE2 gains a 4th comma field "<in>,<out>,<type>,<unitFlags>";
  the old 3-field form still parses (backward compatible).
- Legacy parser migrates the old global bits onto unit 0; KLINKE2
  parser migrates them onto all units.

New HardwareUnit accessors (fakeFaderTouch(), metersOnDisplay(),
switchRows(), needsBlinkEmulation()) replace the global flag reads in
CCSManager (fader/faderTouched/frameUpdate now key off the owning unit).
CSurf_MCU exposes fakeFaderTouch(channel) and anyUnitFakeFaderTouch().

Dialog reorganized into two GROUPBOXes: "Settings for selected unit"
(type, MIDI in/out, the three per-unit checkboxes) and "Global settings
(all units)" (keyboard modifier, swap zoom arrows).

Plug-Mode LCD meter fix (builds on the per-unit metersOnDisplay() flag):
- MeterBridge::showMeterOnDisplay() now respects the per-unit
  metersOnDisplay() option; previously it drew the '|' software-meter
  bars on every unit, overwriting the FADER name/value text on row 1
  (extenders appeared to have no text, and units with the option off
  still showed meters).
- Gate the showMeterOnDisplay() call in updateMeter() on the virtual
  alsoOnDisplay(); PlugModeMeterBridge overrides it to false so Plug
  Mode never draws software-meter bars over its parameter text.
  MultiTrack/Pan/Send keep returning true.
- Add CSurf_MCU::enableMCUMeters(bool, bool excludeProX); PlugMode::
  activate() calls enableMCUMeters(false, /*excludeProX=*/true) so
  Mackie units get the LCD-meter SysEx mode turned off while QCon ProX
  keeps its own hardware meters.

+++ src/core/Options.cpp
@@ -30,7 +30,8 @@ void Options::addAttribute(String optionName, String attribute,
     if (m_optionList[i].first.equalsIgnoreCase(optionName)) {
       m_optionList[i].second.first.push_back(attribute);
       if (defaultAtt) {
-        m_optionList[i].second.second = m_optionList[i].second.first.size() - 1;
+        m_optionList[i].second.second =
+            static_cast<int>(m_optionList[i].second.first.size()) - 1;
       }
     }
   }
@@ -83,7 +84,7 @@ void Options::move(unsigned int index, int steps) {
   } else if (m_optionList[numOpt].second.second >=
              (int)m_optionList[numOpt].second.first.size()) {
     m_optionList[numOpt].second.second =
-        m_optionList[numOpt].second.first.size() - 1;
+        static_cast<int>(m_optionList[numOpt].second.first.size()) - 1;
   }
 
   displaySelectedOptions();
